+++ requirements/static/ci/windows.in
@@ -5,3 +5,4 @@ pygit2<=0.28.2; python_version < '3.8'
 pygit2<1.1.0; python_version == '3.8'
 pygit2>=1.2.0; python_version >= '3.9'
 sed
+pywinrm==0.3.0
