+++ salt/utils/cloud.py
@@ -71,11 +71,23 @@ try:
 except ImportError:
     HAS_PSEXEC = False
 
+
+# Set the minimum version of PyWinrm.
+WINRM_MIN_VER = "0.3.0"
+
+
 try:
     import winrm
     from winrm.exceptions import WinRMTransportError
 
-    HAS_WINRM = True
+    # Verify WinRM 0.3.0 or greater
+    import pkg_resources  # pylint: disable=3rd-party-module-not-gated
+    winrm_pkg = pkg_resources.get_distribution("pywinrm")
+    if not salt.utils.versions.compare(winrm_pkg.version, ">=", WINRM_MIN_VER):
+        HAS_WINRM = False
+    else:
+        HAS_WINRM = True
+
 except ImportError:
     HAS_WINRM = False
 
@@ -1205,7 +1217,8 @@ def deploy_windows(
         opts = {}
 
     if use_winrm and not HAS_WINRM:
-        log.error("WinRM requested but module winrm could not be imported")
+        log.error("WinRM requested but module winrm could not be imported. "
+                  "Ensure you are using version {} or higher.".format(WINRM_MIN_VER))
         return False
 
     starttime = time.mktime(time.localtime())
